pyproject.toml
[tool.poetry]
name = "cryptobot"
version = "0.1.0"
description = ""
authors = ["Raveme491 <michaelsdf491@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.1"
httpx = "^0.23.1"
python-binance = "^1.0.16"
praw = "^7.6.1"
selenium = "^4.7.2"
webdriver-manager = "^3.8.5"
arsenic = "^21.8"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
blue = "^0.9.1"
isort = "^5.11.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
